feat: start

+++ lib/data/models/learning_tracks/chapters_model.dart
@@ -8,7 +8,7 @@ class ChaptersModel {
   String title;
   String videoUrl;
   String videoThumbUrl;
-  int ooz;
+  double ooz;
   String createdAt;
   String updatedAt;
